refactor(player): E.39 — extract resume action
Move the ~165-LOC `resume` action body into `runResume(set, get)`
in `src/store/resumeAction.ts`, following the helper-with-thin-wrapper
pattern from E.37 + E.38.

Covers all three resume branches:
- Orbit guest catch-up to host's live position (same-track → seek;
  different-track → playTrack + deferred seek).
- Radio resume via HTML5 audio.
- Regular track resume: warm engine (audio_resume) or cold start
  (hot-cache promote → getSong refetch → audio_play → seek to
  persisted currentTime, with currentTrack fallback on fetch fail).

Pure code-move. playerStore.ts: 868 → 702 LOC (−166).
